move find_sed_file to synthetic_photometry.py
